Move license authorization tests into rest api test file
Consolidate license tests back into test-components-rest-api.php
using the existing create_test_component helper (direct PHP, no REST),
fixing the "Undefined array key 0" CI failures.

Made-with: Cursor
